hermod-nodejs/porcupine/requirements.txt
numpy
soundfile
paho-mqtt
